Custom charset
Added the custom charset back. From now on, after calling the 'change_charset' subroutine, the new charset is activated. It is stored at the RAM locations $3800 following. The first 64 chars are identical to the standard PETSCII characters, from char nr. 64 on, the chars are custom made symbols, that I've drawn by hand on paper (that was a thing back in the 80s).
The symbols might include minor defects, as I typed them by hand (32 chars; 8 bytes per char; only the original planning with decimal (not hex) on hand; ...).

The next commit will include the main-menu draw-subroutine and the planning ahead. Good night y'all. -J